@@ -3,13 +3,14 @@ package com.degiv.springmvcwebapp.controllers;
 import com.degiv.springmvcwebapp.dao.PersonDAO;
 import com.degiv.springmvcwebapp.models.Person;
 import com.degiv.springmvcwebapp.util.PersonValidator;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
+
 
 @Controller
 @RequestMapping("/people")
